+++ react/packages/@porsche/ui-kit-react/src/components/atoms/Icon/Icon.tsx
@@ -63,11 +63,6 @@ export interface IconProps extends ClassNameProp, ComponentProp {
     onClick?: (event: React.MouseEvent<HTMLElement>, data: IconProps) => void
 
     /**
-     * Puts a circle around the icon
-     */
-    circled?: boolean
-
-    /**
      * Adds a native HTML tooltip to the icon
      */
     title?: string
@@ -114,7 +109,7 @@ const _Icon: React.FunctionComponent<IconProps> & Partial<Icon> = (props) => {
         onClick(e, props)
     }
 
-    const { as, children, size, circled, ...rest } = iconRest
+    const { as, children, size, ...rest } = iconRest
 
     const ElementType = getElementType(as, "i")
     const SVGIcon = (registeredIcons as any)[name] || (defaultIcons as any)[name]
@@ -126,7 +121,6 @@ const _Icon: React.FunctionComponent<IconProps> & Partial<Icon> = (props) => {
     const classes = cx(
         prefix("icon"),
         prefix(`icon--${size}`),
-        { [prefix("icon--circled")]: circled },
         { [prefix(`icon--${actualColor}`)]: actualColor !== undefined },
         { [prefix(`icon--${actualHoverColor}--hover`)]: !disabled && actualHoverColor !== undefined },
         { [prefix(`icon--link`)]: !!onClick && !disabled },